requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
folium>=0.12.0
pyyaml>=5.4.0
branca>=0.4.0